+++ common/src/main/java/uk/gov/justice/services/common/http/HeaderConstants.java
@@ -6,6 +6,7 @@ public final class HeaderConstants {
     public static final String USER_ID = "CJSCPPUID";
     public static final String SESSION_ID = "CPPSID";
     public static final String NAME = "CPPNAME";
+    public static final String CAUSATION = "CPPCAUSATION";
     public static final String CLIENT_CORRELATION_ID = "CPPCLIENTCORRELATIONID";
 
     private HeaderConstants() {